backend/requirements.txt
pg8000
python-dotenv==1.1.1
requests==2.32.5
tzdata
amexcerts==0.1.9
